Like Feature
- Corrections to the likes feature
