+++ src/index.html
@@ -69,10 +69,6 @@
 <footer class="fixed bottom-0 left-0 right-0 py-2 text-center bg-gray-800 text-white">
   <p>Created by Microverse under CC license</p>
 </footer>
-<script type="module" src="moviesData.js"></script>
-<script type="module" src="api.js"></script>
-<script type="module" src="movie.js"></script>
-<script type="module" src="main.js"></script>
 <script src="https://cdn.tailwindcss.com/2.2.4/tailwind.min.js"></script>
 </body>
 </html>
